Fix teams page.
When a user is not logged in, he is neither a coach nor an athlete, so
we should return a Team.objects.none() queryset.

I'm not sure what a public team is since public/private is defined per
workout. I guess we could look through and see which teams have
participated in public sessions, but I'm not going to tackle that now.